requirements.txt
Flask==2.2.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Jinja2==3.1.2
mysql-connector-python==8.0.32
PyMySQL==1.0.2
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
Pyrebase4==4.6.0
pyrebase==3.0.27
Pillow==9.5.0